skills/fabius-cohors/references/agents/python/parallel_task_decomposition_execution/pyproject.toml
[project]
name = "parallel-task-decomposition-execution"
version = "0.1.0"
description = "A sample agent that demonstrates parallel task decomposition and execution."
authors = [
    { name = "Google LLC" }
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "google-adk==1.19.0",
    "python-dotenv==1.2.1",
    "google-cloud-aiplatform[adk,agent-engines]==1.128.0",
    "mcp==1.21.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.7",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.32.0",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
# CRITICAL: Ensures the internal logic is treated as first-party
known-first-party = ["parallel_task_decomposition_execution"]

[tool.hatch.build.targets.wheel]
# Explicitly mapping the package folder to avoid discovery errors
packages = ["parallel_task_decomposition_execution"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties and settings.
[tool.agent-starter-pack]
example_question = "What is the current local time in Sydney, the current weather in New York, and the latest Bitcoin price in USD?"

[tool.agent-starter-pack.settings]
agent_directory = "parallel_task_decomposition_agent"